Move IDE-plugin build helper files to dependencies plugin
