--- kt/buildSrc/build.gradle.kts
@@ -1,16 +1,0 @@
-plugins {
-    `kotlin-dsl`
-}
-
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-}
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation(kotlin("gradle-plugin", version = "1.7.20"))
-}
