+++ kt/plugins/godot-dependencies-gradle-plugin/src/main/kotlin/godot/dependencies/gradle/helper/BuildConfig.kt
@@ -1,4 +1,4 @@
-package plugins.intellij
+package godot.dependencies.gradle.helper
 
 data class BuildConfig(
     val sdk: String,
