+++ kt/plugins/godot-dependencies-gradle-plugin/src/main/kotlin/godot/dependencies/gradle/helper/VersionRange.kt
@@ -1,3 +1,3 @@
-package plugins.intellij
+package godot.dependencies.gradle.helper
 
 data class VersionRange(val since: String, val until: String)
